admin/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
psycopg2==2.8.6
flake8==3.9.2
pika==1.2.0
django-cors-headers